requirements.txt
Flask==3.1.0
setproctitle==1.3.5
speedtest-cli==2.1.3